re-organize chloroform configuration
  default_settings.py is moved under the chloroform package, and all configuration in
  that module is used.  Load settings from env var location, or cfg/settings.py.  Merge
  database.py into chloroform/__init__.py, improved method for starting the server.
