--- cfg/__init__.py
@@ -1,5 +1,0 @@
-'''
-Created on Oct 30, 2016
-
-@author: salami
-'''
