+++ chloroform/__main__.py
@@ -1,0 +1,8 @@
+import sys
+import os
+
+sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
+
+from chloroform import app
+
+app.run()
