+++ samples/build_db.py
@@ -8,7 +8,7 @@ import sys
 sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
 
 from chloroform.models import *
-from chloroform.database import db
+from chloroform import db
 
 db.create_all()
 db.session.commit()
