+++ samples/test_db.py
@@ -4,10 +4,8 @@ import sys
 # adding root of project to sys path
 sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
 
-from chloroform.database import db
+from chloroform import db
 from chloroform.models import *
-
-
 
 client = Client(name = "Simon The Pie Man")
